2to3: Run fixer 'long'

+++ Mopy/bash/bosh/loot_conditions.py
@@ -104,7 +104,7 @@ class ConditionFunc(_ACondition):
         for a in self.func_args:
             if isinstance(a, unicode): # String
                 fmt_a = u'"%s"' % a
-            elif isinstance(a, (int, long)): # Checksum
+            elif isinstance(a, (int, int)): # Checksum
                 fmt_a = u'%X' % a
             else: # Comparison
                 fmt_a = u'%r' % a
